# shared-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Melanie Data API '
  description: 'A high performance & centrally cached API service for premium bots. '
  version: ' commit e2ffc88 ok @ Sep 10 11:37 pm'
servers:
- url: https://dev.melaniebot.net
  description: montreal
paths:
...
  /api/discord/bio:
    get:
      tags:
      - discord
      summary: Fetch Discord Bio
      description: fetch a user's bio & server banner
      operationId: fetchDiscordBio
      parameters:
      - required: true
        schema:
          type: string
          title: User Id
        name: user_id
        in: query
      - required: false
        schema:
          type: string
          title: Guild Id
        name: guild_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BioResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/discord/clearsnipe:
    delete:
      tags:
      - discord
      summary: Issue Clear Snipe
      operationId: issueClearSnipe
      parameters:
      - required: true
        schema:
          type: string
          title: Channel Id
        name: channel_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionConfirmation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/snap/{username}:
    get:
      tags:
      - snapchat
      summary: Get Snapuser
      description: Fetch a snap user!
      operationId: getSnapuser
      parameters:
      - required: true
        schema:
          type: string
          title: Username
        name: username
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BannerType:
      properties:
        banner:
          type: string
          title: Banner
        url:
          type: string
          title: Url
        color:
          $ref: '#/components/schemas/ColorPalette'
        guild_id:
          type: integer
          title: Guild Id
        format:
          type: string
          title: Format
        user_id:
          type: integer
          title: User Id
      type: object
      title: BannerType
    BioMember:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        avatar:
          type: string
          title: Avatar
        avatar_decoration:
          type: string
          title: Avatar Decoration
        discriminator:
          type: string
          title: Discriminator
        public_flags:
          type: integer
          title: Public Flags
        flags:
          type: integer
          title: Flags
        banner:
          $ref: '#/components/schemas/BannerType'
        banner_color:
          type: string
          title: Banner Color
        accent_color:
          type: integer
          title: Accent Color
        bio:
          type: string
          title: Bio
      type: object
      title: BioMember
    BioRequest:
      properties:
        user_id:
          type: integer
          title: User Id
        guild_id:
          type: integer
          title: Guild Id
        sig:
          type: string
          title: Sig
        req_user_id:
          type: integer
          title: Req User Id
        timestamp:
          type: integer
          title: Timestamp
      type: object
      required:
      - user_id
      title: BioRequest
    BioResponse:
      properties:
        user:
          $ref: '#/components/schemas/BioUser'
        profile_data:
          $ref: '#/components/schemas/ProfileModel'
        member:
          $ref: '#/components/schemas/BioMember'
        activities:
          items: {}
          type: array
          title: Activities
        request:
          $ref: '#/components/schemas/BioRequest'
        status:
          $ref: '#/components/schemas/GatewayUserStatus'
        sig:
          type: string
          title: Sig
      type: object
      title: BioResponse
    BioUser:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        avatar:
          type: string
          title: Avatar
        avatar_decoration:
          type: string
          title: Avatar Decoration
        discriminator:
          type: string
          title: Discriminator
        public_flags:
          type: integer
          title: Public Flags
        flags:
          type: integer
          title: Flags
        banner:
          $ref: '#/components/schemas/BannerType'
        banner_color:
          type: string
          title: Banner Color
        accent_color:
          type: integer
          title: Accent Color
        bio:
          type: string
          title: Bio
      type: object
      title: BioUser
...
    ColorPalette:
      properties:
        colors:
          items:
            items:
            - type: integer
            - type: integer
            - type: integer
            type: array
            maxItems: 3
            minItems: 3
          type: array
          title: Colors
      type: object
      required:
      - colors
      title: ColorPalette
    ConnectedAccount:
      properties:
        type:
          type: string
          title: Type
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        verified:
          type: boolean
          title: Verified
      type: object
      title: ConnectedAccount
...
    DeletionConfirmation:
      properties:
        confirmed:
          type: boolean
          title: Confirmed
        confirmed_by:
          items:
            type: string
          type: array
          title: Confirmed By
          default: []
        deleted_items:
          type: integer
          title: Deleted Items
          default: 0
        sig:
          type: string
          title: Sig
      type: object
      required:
      - confirmed
      - sig
      title: DeletionConfirmation
    DiscordUser:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        avatar:
          type: string
          title: Avatar
        avatar_decoration:
          type: string
          title: Avatar Decoration
        discriminator:
          type: string
          title: Discriminator
        public_flags:
          type: integer
          title: Public Flags
        flags:
          type: integer
          title: Flags
        banner:
          type: string
          title: Banner
        banner_color:
          type: string
          title: Banner Color
        accent_color:
          type: integer
          title: Accent Color
        bio:
          type: string
          title: Bio
      type: object
      title: DiscordUser
...
    GatewayUserStatus:
      properties:
        primary:
          type: string
          title: Primary
        desktop:
          type: string
          title: Desktop
        mobile:
          type: string
          title: Mobile
        web:
          type: string
          title: Web
      type: object
      title: GatewayUserStatus
    GuildMember:
      properties:
        flags:
          type: integer
          title: Flags
        is_pending:
          type: boolean
          title: Is Pending
        joined_at:
          type: string
          title: Joined At
        nick:
          type: string
          title: Nick
        pending:
          type: boolean
          title: Pending
        premium_since:
          type: string
          title: Premium Since
        roles:
          items:
            type: string
          type: array
          title: Roles
        user:
          $ref: '#/components/schemas/User1'
        bio:
          type: string
          title: Bio
        banner:
          type: string
          title: Banner
        mute:
          type: boolean
          title: Mute
        deaf:
          type: boolean
          title: Deaf
      type: object
      title: GuildMember
    GuildMemberProfile:
      properties:
        guild_id:
          type: string
          title: Guild Id
        bio:
          type: string
          title: Bio
        banner:
          type: string
          title: Banner
        accent_color:
          type: string
          title: Accent Color
      type: object
      title: GuildMemberProfile
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProfileModel:
      properties:
        user:
          $ref: '#/components/schemas/DiscordUser'
        connected_accounts:
          items:
            $ref: '#/components/schemas/ConnectedAccount'
          type: array
          title: Connected Accounts
        premium_since:
          type: string
          title: Premium Since
        premium_type:
          type: integer
          title: Premium Type
        premium_guild_since:
          type: string
          title: Premium Guild Since
        profile_themes_experiment_bucket:
          type: integer
          title: Profile Themes Experiment Bucket
        guild_member:
          $ref: '#/components/schemas/GuildMember'
        user_profile:
          $ref: '#/components/schemas/UserProfile'
        guild_member_profile:
          $ref: '#/components/schemas/GuildMemberProfile'
      type: object
      title: ProfileModel
...
    SnapProfileResponse:
      properties:
        username:
          type: string
          title: Username
        bio:
          type: string
          title: Bio
        subscriber_count:
          type: integer
          title: Subscriber Count
        profile_image_url:
          type: string
          title: Profile Image Url
        hero_image_url:
          type: string
          title: Hero Image Url
        display_name:
          type: string
          title: Display Name
          description: The in-app name the user has set for their account.
        snapcode_image_url:
          type: string
          title: Snapcode Image Url
          description: Snap QR code.
        bitmoji_url:
          type: string
          title: Bitmoji Url
          description: Bitmoji, rendered as PNG. Transparent background.
        bitmoji_background_url:
          type: string
          title: Bitmoji Background Url
          description: The bitmoji background shown within the Snapchat application
        one_click_url:
          type: string
          title: One Click Url
          description: Direct link that will open the Snapchat app and allow the user
            to be added.
        share_image_url:
          type: string
          title: Share Image Url
          description: A Combo image of the user's bitmoji and snapcode
        story_media:
          items:
            type: string
          type: array
          title: Story Media
          default: []
        spotlight_media:
          items:
            type: string
          type: array
          title: Spotlight Media
          default: []
      type: object
      required:
      - username
      title: SnapProfileResponse
...
    User1:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        avatar:
          type: string
          title: Avatar
        avatar_decoration:
          type: string
          title: Avatar Decoration
        discriminator:
          type: string
          title: Discriminator
        public_flags:
          type: integer
          title: Public Flags
      type: object
      title: User1
...
    UserProfile:
      properties:
        bio:
          type: string
          title: Bio
        accent_color:
          type: integer
          title: Accent Color
        banner:
          type: string
          title: Banner
      type: object
      title: UserProfile
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...